+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 ### :bug: Bug Fix
 - `html`
   - [#338](https://github.com/wix-incubator/rich-content/pull/338) fix performance.now() breaks ssr 
+- `file-upload`
+  - [#340](https://github.com/wix-incubator/rich-content/pull/340) unique ids for icons
 <hr/>
 
 ## 4.0.7 (Jul 24, 2019)